fix bug in schema update where field rename isn't propagated correctly
in this case a field was changed type, and renamed to prevent an error about an
incompatible update. however, the rename of the field wasn't taken into account
when preparing older typeversions for use with the latest type. when later
parsing a record with the older typeversion, a field could be loaded with the
incompatible type, resulting in a reflect panic.