implement indices on slice-type fields
to be used by mox, for efficiently filtering delivered messages based on
dkim-verified domains.

the new query function FilterIn will filter for records where the slice-type
has the specified value. the planner automatically uses an index, just like it
would (with no preference) for FilterEqual with an index on a non-slice field.
a slice-index cannot be "unique". a slice-index can involve multiple fields,
but only max 1 slice field.