add support for cyclic types (but not cyclic data)
when analyzing a type, track numeric references to named types, and when
encountering a type again, just reference it by number. this prevents endless
recursion trying to create a recursive type definition, as we would do before.
this introduces a new "ondiskVersion2".

this required a rewrite of the "nonzero" schema-checking code, because it tried
to enumerate the fields to analyze (before doing any checking) in a way that
didn't work with cyclic types.