add a context parameter to the db/tx functions
this is a backwards incompatible change, but the project is still young, and
untagged. and the fixes should be simple, easiest is to temporarily add
context.TODO() as parameters.

the common case is a context parameter with Open, with DB.Read and DB.Write
(the whole transaction uses the context), and the DB Get/Update/etc and QueryDB
functions (the context is set on the transaction, and propagated).

a Query uses the context from the tx (for QueryTx) or from QueryDB. methods on
Query don't each get a context parameter.  wouldn't be all that useful anyway,
and would make the api less developer-friendly.

this also moves the Types, Keys, Record and Records functions (used for data
export) from DB to Tx. makes more sense to have them there, especially

suggested at a go meetup by nils dijk