bugfix: Query.Delete and probably Query.Update* would not affect all records that match the query
in certain conditions, seemingly based on existence of (surrounding) records,
some records would be skipped by the operation. the bug was that the internal
bbolt Cursor was not repositioned after modifying the bucket, which must be
done.

the fix applied is to first gather all records to delete/update, then make the
changes. this may use more memory.

an alternative solution could be to reposition the cursor after each
delete/update. that solution would be more involved. for updates, that solution
may run into the problem that the cursor scan would be affected by the updated
row.