bugfix: ordering by a part of an index would sort all matching records by the ordering not handled by the index
i.e. the ordering that was supposed to be handled by the index was undone.

we now scan the index, and sort groups of matching records. grouped by matching
by part of the index key covered by the ordering handled by the index. the
sorting we then do is only for the trailing ordering fields not handled by the
index.