while a query is using a cursor, reset the cursor when changes are made to the cursor bucket
The bbolt Cursor type has this warning:

	Changing data while traversing with a cursor may cause it to be invalidated and
	return unexpected keys and/or values. You must reposition your cursor after
	mutating data.

Instead of repositioning our cursor, we ensured Query.Delete and Query.Update*
first collected all keys, and then deleted them. And we added a warning to
Query.ForEach:

	Fn must not update values, the internal cursor is not repositioned between
	invocations of fn, which would cause undefined behaviour (in practice,
	matching values could be skipped).

The downside of collecting all values before deleting is that it may consume quite some memory.
The downside of the warning at ForEach is that it is easily ignored/forgotten, leading to subtle bugs.

So instead we now do reposition the bucket used in cursors if changes are made
to it. During any change to a bucket (index or record), we look into open
queries with a cursor using that bucket, and mark them as needing a seek, which
is handled by the query when it continues.